+++ Rover/GCS_Mavlink.cpp
@@ -150,7 +150,7 @@ void GCS_MAVLINK_Rover::send_servo_out()
         );
 }
 
-int16_t GCS_MAVLINK_Rover::vfr_hud_throttle() const
+float GCS_MAVLINK_Rover::vfr_hud_throttle() const
 {
     return rover.g2.motors.get_throttle();
 }
