+++ Rover/GCS_Mavlink.h
@@ -63,7 +63,7 @@ private:
     MAV_MODE base_mode() const override;
     MAV_STATE vehicle_system_status() const override;
 
-    int16_t vfr_hud_throttle() const override;
+    float vfr_hud_throttle() const override;
 
 #if AP_RANGEFINDER_ENABLED
     void send_rangefinder() const override;
